==> code/10-validating.ipynb <==
# 10. Validating

This script will validate the results that our first data collection got by comapring it to data collections done by other people in other points in time.

import pandas as pd
import json
import glob
import os
import numpy as np
import re

pd.set_option('display.max_rows', 500)

def read_jsons(directory):
    '''
    Reads the JSON files with the Google Image Search results
    and stores them into a single dictionary, which is returned.
    '''
        
    # Grabs all the filepaths
    files = glob.glob(f"{directory}/*.json")
    
    # Reads all those files and save them into a new dictionary
    jsons = { }
    for file in files:
        
        with open(file) as f:
            data = json.load(f)
            
            query = data["search_parameters"]['q']
    
            results = data["images_results"]
                
            jsons[query] = results.copy()
            
    return jsons

def make_df(jsons):
    '''
    Turns a JSON-like array of dictionaries,
    produced by the function read_jsons(),
    into a single pandas datrame, which is returned.
    '''

    dfs = []
    
    # Saves both the values of each dictionary
    # and an identifying column with the respective search query
    for k,v in jsons.items():
        df = pd.DataFrame(v)
        df['search_query'] = k
        dfs.append(df)

    dfs = pd.concat(dfs, ignore_index=True)
    
    return dfs

#### Reads my data in

rm_df = make_df(read_jsons("../output/search_results/"))

## Comparisons

def calculate_overlap(df_a, df_b):
    
    # Creates unique series for images, urls and titles
    a_imgs = pd.Series(df_a["original"].unique())
    b_imgs = pd.Series(df_b["original"].unique())
    
    a_urls = pd.Series(df_a["link"].unique())
    b_urls = pd.Series(df_b["link"].unique())
        
    print("How many of the images in df_a are also in df_b?")
    
    display(
        a_imgs.isin(b_imgs).value_counts(normalize=True)
    )
    
    print("How many of the URLs in df_a are also in df_b?")

    display(
        a_urls.isin(b_urls).value_counts(normalize=True)
    )
    
    print("How many of the images in df_b are also in df_a?")
    
    display(
        b_imgs.isin(a_imgs).value_counts(normalize=True)
    )
    
    print("How many of the URLs in df_b are also in df_a?")

    display(
        b_urls.isin(a_urls).value_counts(normalize=True)
    )

### 1. Gianna Grün

gg_df = make_df(read_jsons("../validation_output/1/search_results/"))

calculate_overlap(rm_df, gg_df)

### 2. Thanasis Troboukis

tt_df = make_df(read_jsons("../validation_output/2/search_results/"))

calculate_overlap(rm_df, tt_df)

### 3. Vinicius Sueiro

vs_df = make_df(read_jsons("../validation_output/3/search_results/"))

calculate_overlap(rm_df, vs_df)

calculate_overlap(tt_df, vs_df)
